Better price handling, use BigDecimal for some pricing math

+++ event-trader/src/main/java/com/jgoetsch/eventtrader/order/price/TradeSignalPrice.java
@@ -28,7 +28,7 @@ public class TradeSignalPrice extends OffsetOrderPrice {
 
 	@Override
 	protected Double getBaseValue(TradeSignal trade, MarketData marketData) {
-		return trade.getPrice();
+		return trade.getPrice().doubleValue();
 	}
 
 }
